update parsing scripts to improve robustness, fix multi-pe advection=none

+++ configuration/scripts/parse_settings.sh
@@ -10,7 +10,7 @@ filename=$1
 filemods=$2
 
 #echo "$0 $1 $2" 
-echo "running parse_settings.sh"
+echo "running ${scriptname}"
 foundstring="FoundSTRING"
 vnamearray=()
 valuearray=()
@@ -23,8 +23,11 @@ do
   else
     #vname=`echo $line | sed "s|\(^\s*set\S*\)\s\{1,100\}\(\S*\)\s\{1,100\}\(\S*\).*$|\2|g"`
     #value=`echo $line | sed "s|\(^\s*set\S*\)\s\{1,100\}\(\S*\)\s\{1,100\}\(\S*\).*$|\3|g"`
-    vname=`echo $line | sed "s|\(^[[:space:]]*set[^[:space:]]*\)[[:space:]][[:space:]]*\([^[:space:]]*\)[[:space:]][[:space:]]*\([^[:space:]]*\).*$|\2|g"`
+    vname=`echo $line | sed "s|\(^[[:space:]]*set[^[:space:]]*\)[[:space:]][[:space:]]*\([^[:space:]]*\).*$|\2|g"`
     value=`echo $line | sed "s|\(^[[:space:]]*set[^[:space:]]*\)[[:space:]][[:space:]]*\([^[:space:]]*\)[[:space:]][[:space:]]*\([^[:space:]]*\).*$|\3|g"`
+    if [[ "${value}" == "${line}" ]]; then
+       value=""
+    fi
 #    echo "$line $vname $value"
 
     found=${foundstring}
@@ -44,18 +47,25 @@ do
     done
 
     #sed -i 's|\(^\s*set.* '"$vname"' \)[^#]*\(#*.*$\)|\1 '"$value"'  \2|g' $filename
-    sed -i.sedbak -e 's|\(^[[:space:]]*set.* '"$vname"' \)[^#]*\(#*.*$\)|\1 '"$value"'  \2|g' $filename
-
-    if [[ "${found}" == "${foundstring}" ]]; then
-      vnamearray+=($vname)
-      valuearray+=($value)
+    cp ${filename} ${filename}.check
+    sed -i.sedbak -e 's|\(^[[:space:]]*set.* '"$vname"' \)[^#]*\(#*.*$\)|\1 '"$foundstring"' |g' $filename.check
+    grep -q ${foundstring} ${filename}.check
+    if [ $? -eq 0 ]; then
+      sed -i.sedbak -e 's|\(^[[:space:]]*set.* '"$vname"' \)[^#]*\(#*.*$\)|\1 '"$value"'  \2|g' $filename
+      if [[ "${found}" == "${foundstring}" ]]; then
+        vnamearray+=($vname)
+        valuearray+=($value)
+      else
+        valuearray[$found]=${value}
+      fi
+      if [[ -e "${filename}.sedbak" ]]; then
+        rm ${filename}.sedbak
+      fi
     else
-      valuearray[$found]=${value}
+      echo "${scriptname} ERROR: parsing error for ${vname}"
+      exit -99
     fi
-
-    if [[ -e "${filename}.sedbak" ]]; then
-      rm ${filename}.sedbak
-    fi
+    rm ${filename}.check ${filename}.check.sedbak
 
   fi
